+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/dungeons/AbstractDungeon/FixLogging.java
@@ -1,10 +1,5 @@
 package basemod.patches.com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 
-import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
-import javassist.CannotCompileException;
-import javassist.expr.ExprEditor;
-import javassist.expr.MethodCall;
-
 public class FixLogging
 {
     //Makes it possible to prefix/postfix patch these methods instead of making the default branch almost impossible to patch
@@ -15,41 +10,7 @@ public class FixLogging
 
     }
 
-    @SpirePatch(
-            cls="com.megacrit.cardcrawl.dungeons.AbstractDungeon",
-            method="update"
-    )
-    public static class FixUpdateLog {
-        public static ExprEditor Instrument()
-        {
-            return new ExprEditor() {
-                @Override
-                public void edit(MethodCall m) throws CannotCompileException
-                {
-                    if (m.getMethodName().equals("info")) {
-                        m.replace(FixLogging.class.getName() + ".unknownScreenUpdate();");
-                    }
-                }
-            };
-        }
-    }
+    public static void unknownScreenClose() {
 
-    @SpirePatch(
-            cls="com.megacrit.cardcrawl.dungeons.AbstractDungeon",
-            method="render"
-    )
-    public static class FixRenderLog {
-        public static ExprEditor Instrument()
-        {
-            return new ExprEditor() {
-                @Override
-                public void edit(MethodCall m) throws CannotCompileException
-                {
-                    if (m.getMethodName().equals("info")) {
-                        m.replace(FixLogging.class.getName() + ".unknownScreenRender();");
-                    }
-                }
-            };
-        }
     }
 }